Implement the first major bit of network code as a proof of concept.
This is mostly as a jumping off point to doing the real interesting stuff.  A lot of this has been brutally hacked together from ry's node_postgres example and the mongo c driver sources (which involved some not so gentle copy-and-pasting. At the present moment, find doesn't take any arguments, and is hardcoded to a particular db. OID's aren't returned yet. There is only a single object type, Connection, that is a wrapper around the Connection and returned cursors.